:recycle: reorganize files
