+++ src/const/strings.js
